feat(HRME-30): backlog categorya subs
